vimrc: Added a mapping from Ctrl+[ to Esc.
This is to work around the fact that PCKeyboardHack doesn't work yet in
Lion, and I had been using that to map Caps Lock to Escape. But, System
Preferences allows you to map Caps Lock to Control, so I can do that and
then use this mapping without being reliant on any external tool. Thanks
to Tony for the idea.